Add Dockerfile and requirements for backend; implement Neo4j Aura config loader
- Created Dockerfile for backend setup, including dependencies for Flask API and OSV ingestor.
- Added requirements.txt for backend with Flask, Neo4j, requests, and tqdm.
- Implemented neo4j_aura_config.py to manage Neo4j Aura credentials loading from environment or TXT files.
- Introduced repo_progress.json to track processed repositories and current batch size.
- Added requirements.txt for VulGPT_OSV with necessary libraries.
