+++ RepairGPT/Dockerfile
@@ -1,11 +1,9 @@
-FROM python:3.9-slim
+FROM python:3.11-slim
 
 WORKDIR /app
 
-# Copy requirements.txt (create if it doesn't exist)
-RUN echo "Flask==2.0.1" > requirements.txt
-
 # Install dependencies
+COPY requirements.txt ./requirements.txt
 RUN pip install --no-cache-dir -r requirements.txt
 
 # Copy source code
