+++ RepairGPT/ollama_app.py
@@ -4,15 +4,28 @@ import requests
 import time
 import argparse
 import os
+import sys
 import datetime
 import re
 import logging
 from collections import Counter
 from typing import List, Dict, Optional, Any, Tuple
 from dataclasses import dataclass
+from pathlib import Path
 
 from neo4j import GraphDatabase
 from neo4j.exceptions import ServiceUnavailable, AuthError
+
+# Attempt to auto-load Neo4j Aura credentials from the repo-root TXT file.
+try:
+    REPO_ROOT = Path(__file__).resolve().parents[1]
+    if str(REPO_ROOT) not in sys.path:
+        sys.path.insert(0, str(REPO_ROOT))
+    from neo4j_aura_config import ensure_neo4j_env_loaded
+
+    ensure_neo4j_env_loaded()
+except Exception:
+    pass
 
 @dataclass
 class VulnerabilityInfo:
@@ -322,9 +335,10 @@ class SecurityAnalyzer:
 class Neo4jSecurityAnalyzer:
     def __init__(
         self,
-        neo4j_uri: str = "bolt://localhost:7687",
-        neo4j_user: str = "neo4j",
-        neo4j_password: str = "jaguarai",
+        neo4j_uri: Optional[str] = None,
+        neo4j_user: Optional[str] = None,
+        neo4j_password: Optional[str] = None,
+        neo4j_database: Optional[str] = None,
         log_level: str = "INFO"
     ):
         """Initialize the security analyzer with connection parameters."""
@@ -338,6 +352,18 @@ class Neo4jSecurityAnalyzer:
             format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
         )
         self.logger = logging.getLogger(__name__)
+
+        neo4j_uri = neo4j_uri or os.environ.get("NEO4J_URI") or os.environ.get("VITE_NEO4J_URI") or "neo4j://localhost:7687"
+        neo4j_user = (
+            neo4j_user
+            or os.environ.get("NEO4J_USERNAME")
+            or os.environ.get("NEO4J_USER")
+            or os.environ.get("VITE_NEO4J_USER")
+            or "neo4j"
+        )
+        neo4j_password = neo4j_password or os.environ.get("NEO4J_PASSWORD") or os.environ.get("VITE_NEO4J_PASSWORD") or ""
+        neo4j_database = neo4j_database or os.environ.get("NEO4J_DATABASE") or os.environ.get("VITE_NEO4J_DATABASE")
+        self.neo4j_database = neo4j_database
         
         # Initialize Neo4j connection
         try:
@@ -357,7 +383,12 @@ class Neo4jSecurityAnalyzer:
     def query_neo4j(self, query: str, params: Dict = None) -> List[Dict]:
         """Execute a Cypher query against Neo4j."""
         try:
-            with self.driver.session() as session:
+            if self.neo4j_database:
+                session_ctx = self.driver.session(database=self.neo4j_database)
+            else:
+                session_ctx = self.driver.session()
+
+            with session_ctx as session:
                 result = session.run(query, params or {})
                 return result.data()
         except Exception as e:
@@ -1318,18 +1349,31 @@ class CombinedSecurityAnalyzer:
     """
     def __init__(
         self,
-        neo4j_uri: str = "bolt://localhost:7687",
-        neo4j_user: str = "neo4j",
-        neo4j_password: str = "jaguarai",
+        neo4j_uri: Optional[str] = None,
+        neo4j_user: Optional[str] = None,
+        neo4j_password: Optional[str] = None,
+        neo4j_database: Optional[str] = None,
         ollama_base_url: str = "http://localhost:11434",
         api_key: str = None,
         log_level: str = "INFO"
     ):
+        neo4j_uri = neo4j_uri or os.environ.get("NEO4J_URI") or os.environ.get("VITE_NEO4J_URI") or "neo4j://localhost:7687"
+        neo4j_user = (
+            neo4j_user
+            or os.environ.get("NEO4J_USERNAME")
+            or os.environ.get("NEO4J_USER")
+            or os.environ.get("VITE_NEO4J_USER")
+            or "neo4j"
+        )
+        neo4j_password = neo4j_password or os.environ.get("NEO4J_PASSWORD") or os.environ.get("VITE_NEO4J_PASSWORD") or ""
+        neo4j_database = neo4j_database or os.environ.get("NEO4J_DATABASE") or os.environ.get("VITE_NEO4J_DATABASE")
+
         # Initialize both analyzers
         self.neo4j_analyzer = Neo4jSecurityAnalyzer(
             neo4j_uri=neo4j_uri,
             neo4j_user=neo4j_user,
             neo4j_password=neo4j_password,
+            neo4j_database=neo4j_database,
             log_level=log_level
         )
         
@@ -1495,9 +1539,26 @@ def main():
     parser.add_argument("--log-level", default="INFO", choices=["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"], help="Logging level")
     
     # Neo4j connection arguments
-    parser.add_argument("--neo4j-uri", default="bolt://localhost:7687", help="Neo4j connection URI")
-    parser.add_argument("--neo4j-user", default="neo4j", help="Neo4j username")
-    parser.add_argument("--neo4j-password", default="jaguarai", help="Neo4j password")
+    parser.add_argument(
+        "--neo4j-uri",
+        default=os.environ.get("NEO4J_URI") or os.environ.get("VITE_NEO4J_URI") or "neo4j://localhost:7687",
+        help="Neo4j connection URI",
+    )
+    parser.add_argument(
+        "--neo4j-user",
+        default=os.environ.get("NEO4J_USERNAME") or os.environ.get("NEO4J_USER") or os.environ.get("VITE_NEO4J_USER") or "neo4j",
+        help="Neo4j username",
+    )
+    parser.add_argument(
+        "--neo4j-password",
+        default=os.environ.get("NEO4J_PASSWORD") or os.environ.get("VITE_NEO4J_PASSWORD") or "",
+        help="Neo4j password",
+    )
+    parser.add_argument(
+        "--neo4j-database",
+        default=os.environ.get("NEO4J_DATABASE") or os.environ.get("VITE_NEO4J_DATABASE"),
+        help="Neo4j database name (Aura often requires this)",
+    )
     
     # Ollama arguments
     parser.add_argument("--ollama-url", default="http://localhost:11434", help="Base URL for Ollama API")
@@ -1660,6 +1721,7 @@ def main():
         neo4j_uri=args.neo4j_uri,
         neo4j_user=args.neo4j_user,
         neo4j_password=args.neo4j_password,
+        neo4j_database=args.neo4j_database,
         log_level=args.log_level
     )
