+++ VulGPT_OSV/repo_progress.json
@@ -1,0 +1,1 @@
+{"processed_repos": ["https://github.com/rails/jquery-rails", "https://github.com/omarkurt/django-connector-CVE-2025-64459-testbed", "https://github.com/expressjs/express", "https://github.com/django/daphne", "https://github.com/jquery/jquery", "https://github.com/django/django", "https://github.com/github/advisory-database", "https://github.com/pypa/advisory-database", "https://github.com/axios/axios", "https://github.com/maximebf/php-debugbar", "https://github.com/lodash/lodash", "https://github.com/pypa/advisory-db", "https://github.com/koajs/koa", "https://github.com/pillarjs/encodeurl", "https://github.com/rubysec/ruby-advisory-db"], "current_batch_size": 6, "timestamp": "2026-03-24T14:52:21.569570"}
