+++ VulGPT_OSV/requirements.txt
@@ -1,0 +1,3 @@
+neo4j==6.1.0
+requests==2.32.5
+tqdm==4.67.3
