+++ backend/requirements.txt
@@ -1,0 +1,4 @@
+Flask==3.0.2
+neo4j==5.23.1
+requests==2.32.5
+tqdm==4.67.3
