+++ docker-compose.yml
@@ -18,37 +18,36 @@ services:
   # RepairGPT service
   repairgpt:
     build:
-      context: ./RepairGPT
-      dockerfile: Dockerfile
+      context: .
+      dockerfile: backend/Dockerfile
     restart: unless-stopped
+    env_file:
+      - ${NEO4J_ENV_FILE:-./Neo4j-6522133f-Created-2026-03-16.txt}
     networks:
       - awesome-iguana-network
-    volumes:
-      - ./RepairGPT:/app
-      - node_modules_repairgpt:/app/node_modules
 
   # VulGPT_OSV service
   vulgpt:
     build:
-      context: ./VulGPT_OSV
-      dockerfile: Dockerfile
+      context: .
+      dockerfile: backend/Dockerfile
     restart: unless-stopped
+    env_file:
+      - ${NEO4J_ENV_FILE:-./Neo4j-6522133f-Created-2026-03-16.txt}
     networks:
       - awesome-iguana-network
-    volumes:
-      - ./VulGPT_OSV:/app
-      - node_modules_vulgpt:/app/node_modules
-    
-  # Database for the application
-  db:
-    image: postgres:14
+    command: ["python", "vulgpt_osv/osv.py"]
+
+  # Ollama (local LLM runtime)
+  ollama:
+    image: ollama/ollama:latest
     restart: unless-stopped
-    volumes:
-      - postgres_data:/var/lib/postgresql/data
+    ports:
+      - "11435:11434"
     environment:
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=postgres
-      - POSTGRES_DB=iguanas
+      - OLLAMA_HOST=0.0.0.0:11434
+    volumes:
+      - ollama_data:/root/.ollama
     networks:
       - awesome-iguana-network
 
@@ -71,9 +70,7 @@ networks:
     driver: bridge
 
 volumes:
-  postgres_data:
   caddy_data:
   caddy_config:
+  ollama_data:
   node_modules_frontend:
-  node_modules_repairgpt:
-  node_modules_vulgpt:
