+++ frontend/.env.production
@@ -1,14 +1,13 @@
-# Production environment (GCP)
-VITE_NEO4J_URI=bolt://localhost:7687
-VITE_NEO4J_USER=neo4j
-VITE_NEO4J_PASSWORD=jaguarai 
-
-# Ollama API Configuration
-VITE_OLLAMA_API_URL=http://localhost:11434
-VITE_OLLAMA_MODEL=deepseek-r1:7b
-
-# Github Token
-VITE_GITHUB_TOKEN=put token here
-
-#NVD API key can be created here https://nvd.nist.gov/developers/request-an-api-key
-VITE_NIST_API_KEY=your-own-key-here
+# Production environment
+# NOTE: Vite does NOT support indirection like `.env.local.VAR_NAME`.
+# Values in env files must be literal values.
+#
+# In production, provide these at build-time (or create an uncommitted `frontend/.env.local` for local dev):
+# - VITE_NEO4J_URI=neo4j+s://<id>.databases.neo4j.io
+# - VITE_NEO4J_USER=<username>
+# - VITE_NEO4J_PASSWORD=<password>
+# - VITE_NEO4J_DATABASE=<optional database name>
+#
+# Optional Ollama settings for features that call Ollama directly from the browser:
+# - VITE_OLLAMA_API_URL=http://localhost:11434
+# - VITE_OLLAMA_MODEL=deepseek-r1:7b
