+++ frontend/.gitignore
@@ -30,3 +30,6 @@ dist-ssr
 
 # package-lock.json
 *package-lock.json
+
+# repo progress
+*repo_progress.json
