+++ frontend/src/services/neo4j/llmVulnerabilityEvaluator.js
@@ -1,8 +1,9 @@
 import { apiClient } from './axiosConfig';
 
 export class LLMVulnerabilityEvaluator {
-  constructor(driver) {
+  constructor(driver, database) {
     this.driver = driver;
+    this.database = database || import.meta.env.VITE_NEO4J_DATABASE;
     this.ollamaApiUrl = import.meta.env.VITE_OLLAMA_API_URL || 'http://localhost:11434';
     this.ollamaModel = import.meta.env.VITE_OLLAMA_MODEL || 'llama2:latest';
     
@@ -17,6 +18,18 @@ export class LLMVulnerabilityEvaluator {
     
     // Allowed extensions for GitHub API file fetching
     this.allowedExtensions = ['.js', '.py', '.java', '.c', '.cpp', '.h', '.hpp', '.cs', '.rb', '.php', '.ts', '.go', '.rs'];
+  }
+
+  createSession() {
+    if (!this.driver) {
+      throw new Error('Neo4j driver is not initialized');
+    }
+
+    if (this.database) {
+      return this.driver.session({ database: this.database });
+    }
+
+    return this.driver.session();
   }
   
   /**
@@ -94,7 +107,7 @@ export class LLMVulnerabilityEvaluator {
    * Applies filtering rules based on requirements
    */  
   async getCodeFilesForRevision(repoId, versionId) {
-    const session = this.driver.session();
+    const session = this.createSession();
     try {
       const result = await session.run(`
         MATCH (r:Repository {id: $repoId})-[:HAS_VERSION]->(v:Version {id: $versionId})-[:CONTAINS]->(f:File)
@@ -129,7 +142,7 @@ export class LLMVulnerabilityEvaluator {
     console.log(`Starting code file retrieval for repository ID: ${repoId}`);
     
     // First, try to get files from Neo4j
-    const session = this.driver.session();
+    const session = this.createSession();
     try {
       // Try multiple relationship patterns to find files
       // This handles different database schemas that might exist
@@ -322,7 +335,7 @@ export class LLMVulnerabilityEvaluator {
    * Gets relevant vulnerabilities for a repository version
    */
   async getRelevantVulnerabilities(repoId, versionId) {
-    const session = this.driver.session();
+    const session = this.createSession();
     try {
       const result = await session.run(`
         MATCH (r:Repository {id: $repoId})-[:HAS_VERSION]->(v:Version {id: $versionId})
@@ -350,7 +363,7 @@ export class LLMVulnerabilityEvaluator {
    * Gets relevant vulnerabilities for a repository without needing a specific version
    */
   async getRelevantVulnerabilitiesForRepo(repoId) {
-    const session = this.driver.session();
+    const session = this.createSession();
     try {
       const result = await session.run(`
         MATCH (r:Repository {id: $repoId})-[:USES_PACKAGE]->(p:Package)<-[:AFFECTS]-(vuln:Vulnerability)
@@ -587,7 +600,7 @@ Be precise and concise. Base your classification strictly on the evidence in the
    * Stores evaluation results in the database
    */
   async storeEvaluationResult(repoId, versionId, vulnId, evaluationResult) {
-    const session = this.driver.session();
+    const session = this.createSession();
     try {
       await session.run(`
         MATCH (r:Repository {id: $repoId})-[:HAS_VERSION]->(v:Version {id: $versionId})
@@ -625,7 +638,7 @@ Be precise and concise. Base your classification strictly on the evidence in the
    * Stores evaluation results directly on the repository node (for repositories without versions)
    */
   async storeRepositoryEvaluationResult(repoId, vulnId, evaluationResult) {
-    const session = this.driver.session();
+    const session = this.createSession();
     try {
       await session.run(`
         MATCH (r:Repository {id: $repoId})
@@ -662,7 +675,7 @@ Be precise and concise. Base your classification strictly on the evidence in the
    * Calculate evaluation metrics for LLM performance (bonus feature)
    */
   async calculateEvaluationMetrics() {
-    const session = this.driver.session();
+    const session = this.createSession();
     try {
       // Get all evaluations
       const result = await session.run(`
