+++ frontend/src/services/neo4j/vulnerabilityProcessor.js
@@ -43,7 +43,7 @@ function extractSeverity(vuln) {
 export async function storeVulnerability(vulnerability) {
   if (!vulnerability) return;
   
-  const session = this.driver.session();
+  const session = this.createSession();
   try {
     // Create vulnerability node with all fields
     await session.run(`
